load_data_from_bigquery/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery